requirements.txt
tensorflow==2.12.0
opencv-python==4.8.0.76
mediapipe==0.10.21
PyQt5==5.15.9
numpy>=1.22.0,<1.24.0
scikit-learn==1.3.0
pillow>=8.0.0
matplotlib>=3.0.0
h5py>=2.9.0 